カスタムフックとEither型で盤面に関連する処理を切り出した
